started adding more unit tests
