+++ tests/integration/test_tracking.py
@@ -3,49 +3,55 @@ import unittest
 from tempfile import TemporaryDirectory
 from deepdiff import DeepDiff 
 
-from extractor import quadrilaterals
-from tests.common import temp_dir_prefix, load_file
+from extractor import tracking
+from .common import temp_dir_prefix, load_file
 
 
-class TestQuadrilaterals(unittest.TestCase):
+class TestTracking(unittest.TestCase):
 
     def setUp(self):
         self.data_dir = os.path.join("tests", "data", "large")
         self.work_dir = TemporaryDirectory(prefix=temp_dir_prefix)
         self.settings = {
-            "min_iou": 0.9
+            "motion_model": "homography",
+            "orb_nfeatures": 5000,
+            "orb_fast_thres": 12,
+            "orb_scale_factor": 1.2,
+            "orb_nlevels": 8,
+            "match_distance_thres": 20.0,
+            "max_distance": 60,
+            "output_video_fps": 8.0,
+            "deterministic_track_ids": True
         }
         self.frames_root = os.path.join(self.data_dir, "splitted")
         self.inference_root = os.path.join(self.data_dir, "segmented")
-        self.tracks_root = os.path.join(self.data_dir, "tracking")
         self.output_dir = self.work_dir.name
 
         # where to load files with desired output format from
-        self.ground_truth_dir = os.path.join(self.data_dir, "quadrilaterals")
+        self.ground_truth_dir = os.path.join(self.data_dir, "tracking")
 
     def test_run(self):
-        quadrilaterals.run(
+        tracking.run(
             self.frames_root, 
             self.inference_root, 
-            self.tracks_root,
-            self.output_dir, 
+            self.output_dir,
             **self.settings)
 
         # check if outputs equal ground truth
-        file_name = "quadrilaterals.pkl"
+        file_name = "tracks.csv"
         content, content_ground_truth = load_file(
             self.output_dir, 
             self.ground_truth_dir, 
             file_name)
 
+        # Note: Test will fail when order of rows in tracks is permutated
         self.assertEqual(
             DeepDiff(
                 content_ground_truth, 
-                content,
-                math_epsilon=1e-5
+                content
             ), {},
             "{} differs from ground truth".format(file_name)
-        )
+        )        
 
     def tearDown(self):
         self.work_dir.cleanup()
